+++ CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Renamed `Tool.link_name` to `Tool.connected_to`
 * Migrate to COMPAS 2.x: add dependency to `compas_robots`
 * Migrate to COMPAS 2.x: use `compas.tolerance` module instead of `compas.PRECISION`
-
+* Changed base class of `Trajectory` to be `compas.datastructures.Datastructure`, which adds `attributes` to it.
 
 ### Removed